pyproject.toml
[project]
name = "langchainlearn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=7.1.0",
    "langchain>=1.2.6",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.0",
    "langchain-huggingface>=1.2.0",
    "langchain-tavily>=0.2.17",
    "langchain-text-splitters>=1.1.0",
    "pymilvus[milvus-lite]>=2.6.6",
    "python-dotenv>=1.2.1",
    "unstructured[docx]>=0.18.27",
]
